rofs: do not forward permissions from cache when file are already open
When a file is open, a loopback handler is returned and attribute from
local cache is used for underlaying call.

This cause mismatch attributes (permissions, owner, ...) as expected
when a file is already open and attributes are requested.

This fix this behavior. A copy of expected attributes are saved during
the Open call request and theses attributes are returned for underlaying
calls.